refactor: move query building logic from controller to opensearch nn seeker
